+++ src/app/pages/ArticlePage/ArticlePage.tsx
@@ -35,13 +35,11 @@ import CpsRecommendations from '#containers/CpsRecommendations';
 import InlinePodcastPromo from '#containers/PodcastPromo/Inline';
 import {
   Article,
-  OptimoBlock,
   OptimoBylineBlock,
   OptimoBylineContributorBlock,
   Recommendation,
 } from '#app/models/types/optimo';
 import ScrollablePromo from '#components/ScrollablePromo';
-import { MostReadData } from '#app/components/MostRead/types';
 import ElectionBanner from './ElectionBanner';
 import ImageWithCaption from '../../components/ImageWithCaption';
 import AdContainer from '../../components/Ad';
@@ -66,6 +64,7 @@ import Disclaimer from '../../components/Disclaimer';
 import SecondaryColumn from './SecondaryColumn';
 import styles from './ArticlePage.styles';
 import { ComponentToRenderProps, TimeStampProps } from './types';
+import { transformRecsData } from './recommendationsExperiment';
 
 const getImageComponent =
   (preloadLeadImageToggle: boolean) => (props: ComponentToRenderProps) => (
@@ -112,87 +111,6 @@ const DisclaimerWithPaddingOverride = (props: ComponentToRenderProps) => (
 
 const getPodcastPromoComponent = (podcastPromoEnabled: boolean) => () =>
   podcastPromoEnabled ? <InlinePodcastPromo /> : null;
-
-export type Variation = 'wsoj' | 'mostRead' | 'relatedContent';
-
-type TransformRecsDataProps = {
-  wsojRecs: Recommendation[];
-  mostRead: MostReadData;
-  pageBlocks: OptimoBlock[];
-  variation: Variation;
-};
-
-const transformRecsData = ({
-  wsojRecs,
-  mostRead,
-  pageBlocks,
-  variation,
-}: TransformRecsDataProps) => {
-  if (variation === 'wsoj') return wsojRecs;
-
-  if (variation === 'relatedContent') {
-    const relatedContentBlock = pageBlocks.find(
-      block => block.type === 'relatedContent',
-    );
-
-    if (!relatedContentBlock) return null;
-
-    // @ts-expect-error - nested block structure
-    const relatedContentItems = relatedContentBlock?.model?.blocks?.slice(0, 4);
-
-    const transformedRelatedContentItems = relatedContentItems?.map(
-      (item: OptimoBlock) => {
-        // @ts-expect-error - nested block structure
-        const itemBlocks = item?.model?.blocks;
-
-        const imageBlock = itemBlocks?.find(
-          (block: OptimoBlock) => block.type === 'image',
-        );
-        const textBlock = itemBlocks?.find(
-          (block: OptimoBlock) => block.type === 'text',
-        );
-        const canonicalUrl =
-          textBlock?.model?.blocks?.[0]?.model?.blocks?.[0]?.model?.locator;
-
-        return {
-          headlines: {
-            promoHeadline: {
-              blocks: textBlock?.model?.blocks,
-            },
-          },
-          images: {
-            defaultPromoImage: {
-              blocks: imageBlock?.model?.blocks,
-            },
-          },
-          locators: { canonicalUrl },
-        };
-      },
-    );
-
-    return transformedRelatedContentItems;
-  }
-
-  if (variation === 'mostRead') {
-    const mostReadItems = mostRead?.items?.slice(0, 4);
-
-    const transformedMostReadItems = mostReadItems?.map(item => {
-      return {
-        headlines: {
-          promoHeadline: {
-            blocks: [singleTextBlock(item.title)],
-          },
-        },
-        images: item?.images,
-        locators: { canonicalUrl: item?.href },
-      };
-    });
-
-    return transformedMostReadItems;
-  }
-
-  return null;
-};
 
 const ArticlePage = ({ pageData }: { pageData: Article }) => {
   const { isApp } = useContext(RequestContext);
@@ -250,7 +168,7 @@ const ArticlePage = ({ pageData }: { pageData: Article }) => {
     wsojRecs: recommendationsData,
     mostRead: mostReadInitialData,
     pageBlocks: blocks,
-    variation: 'mostRead',
+    variation: 'related_content',
   });
 
   const isPGL = pageData?.metadata?.type === PHOTO_GALLERY_PAGE;
